+++ api/accounts/admin.py
@@ -2,6 +2,7 @@ from authemail.admin import (EmailUserAdmin, PasswordResetCodeInline,
                              SignupCodeInline)
 from django.contrib import admin
 from django.contrib.auth import get_user_model
+from django.utils.translation import gettext_lazy as _
 
 
 class MyPasswordResetCodeInline(PasswordResetCodeInline):
@@ -15,12 +16,12 @@ class MySignupCodeInline(SignupCodeInline):
 class MyUserAdmin(EmailUserAdmin):
     fieldsets = (
         (None, {'fields': ('email', 'password')}),
-        ('Personal Info', {'fields': ('first_name', 'last_name')}),
-        ('Permissions', {'fields': ('is_active', 'is_staff',
-                                    'is_superuser', 'is_verified',
-                                    'groups', 'user_permissions')}),
-        ('Important dates', {'fields': ('last_login', 'date_joined')}),
-        ('Custom info', {'fields': ('phone', 'receive_notifications',)}),
+        (_('Personal Info'), {'fields': ('first_name', 'last_name')}),
+        (_('Permissions'), {'fields': ('is_active', 'is_staff',
+                                       'is_superuser', 'is_verified',
+                                       'groups', 'user_permissions')}),
+        (_('Important dates'), {'fields': ('last_login', 'date_joined')}),
+        (_('Custom info'), {'fields': ('phone', 'receive_notifications',)}),
     )
     inlines = [MyPasswordResetCodeInline, MySignupCodeInline]