+++ api/team/admin.py
@@ -1,16 +1,19 @@
 from django.contrib import admin
+from django.utils.translation import gettext_lazy as _
 
 from .models import Team
 
 
 class TeamAdmin(admin.ModelAdmin):
-    list_display = ('name', 'beachname',)
+    list_display = ('name', 'beachname', 'tournament', 'trainer', 'paid', 'state')
+    search_fields = ('name', 'beachname', 'tournament', 'trainer')
+    list_filter = ('paid', 'state', 'tournament__name')
     fieldsets = (
         (None, {'fields': ('name', 'beachname',)}),
-        ('Date', {'fields': ('date_signup',)}),
-        ('Status', {'fields': ('state',
-                               'paid',)}),
-        ('Relations', {'fields': ('trainer', 'tournament',)})
+        (_('Date'), {'fields': ('date_signup',)}),
+        (_('Status'), {'fields': ('state',
+                                  'paid',)}),
+        (_('Relations'), {'fields': ('trainer', 'tournament',)})
     )